+++ Source/CardGame/Classes/Services/CardGameAttributeService.h
@@ -9,8 +9,8 @@ struct FCardGameScopedModel;
 class CARDGAME_API FCardGameAttributeService
 {
 public:
-	float GetAttributeValue(const FCardGameScopedModel& ScopedModel, UCardGameAttribute* Attribute) const;
-	void SetAttributeValue(FCardGameScopedModel& ScopedModel, UCardGameAttribute* Attribute, float NewValue) const;
-	float GetGlobalAttributeValue(const FCardGameModel& Model, UCardGameAttribute* Attribute) const;
-	void SetGlobalAttributeValue(FCardGameModel& Model, UCardGameAttribute* Attribute, float NewValue) const;
+	int32 GetAttributeValue(const FCardGameScopedModel& ScopedModel, UCardGameAttribute* Attribute) const;
+	void SetAttributeValue(FCardGameScopedModel& ScopedModel, UCardGameAttribute* Attribute, int32 NewValue) const;
+	int32 GetGlobalAttributeValue(const FCardGameModel& Model, UCardGameAttribute* Attribute) const;
+	void SetGlobalAttributeValue(FCardGameModel& Model, UCardGameAttribute* Attribute, int32 NewValue) const;
 };
